+++ CCApi.Extensions.DependencyInjection/Models/PagedResponseDto.cs
@@ -1,0 +1,15 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace CCApi.Extensions.DependencyInjection.Models
+{
+    public class PagedResponseDto<T>
+    {
+        public int Total { get; set; }
+        public int Offset { get; set; }
+        public int Limit { get; set; }
+        public List<T> Data { get; set; }
+    }
+}
